Updating and renaming begin_round.js
Updating begin_round.js by adding a new function and renaming it to. You may think: "Oh, no! Steven is commenting out all the code!". And as a matter of fact that is true! I am commiting this to a new branch and also will soon be adding new functions to make this work again.